refactor : moved 'AccountInfoInterface' into common types folder, renamed auth folder - reason of moved 'AccountInfoInterface' into common types folder : interface is using in another file too - reason of renamed auth folder : it is convention to use auth/[...nextauth]
