+++ src/apis/authentication.ts
@@ -1,10 +1,5 @@
 import { API_AUTHENTICATION_URL } from '@/constants/common'
-
-interface AccountInfoInterface {
-  name: string
-  email: string
-  password: string
-}
+import { AccountInfoInterface } from '@/types'
 
 export const registerAccount = async (accountInfo: AccountInfoInterface) => {
   const res = await fetch(`${API_AUTHENTICATION_URL}/register`, {
