+++ src/app/api/auth/userExists/route.ts
