+++ src/types/account.ts
@@ -1,0 +1,5 @@
+export interface AccountInfoInterface {
+  name: string
+  email: string
+  password: string
+}
